LMstudio_bot/requirements.txt
scikit-learn>=1.0.2
numpy>=1.21.0
discord.py>=2.0.0
python-dotenv>=0.19.0
requests>=2.26.0
openai>=0.27.0
aiohttp>=3.8.0
openai>=1.12.0  # Added for OpenAI API compatibility
certifi>=2024.2.2  # Added for SSL certificate validation